+++ cmd/tempo/app/app.go
@@ -345,8 +345,6 @@ func (t *App) readyHandler(sm *services.Manager, shutdownRequested *atomic.Bool)
 			return
 		}
 
-		// Generator has a special check that makes sure that it was able to register into the ring,
-		// and that all other ring entries are OK too.
 		if t.generator != nil {
 			if err := t.generator.CheckReady(r.Context()); err != nil {
 				http.Error(w, "Generator not ready: "+err.Error(), http.StatusServiceUnavailable)